refactor: extract shared PluginIcon component, remove ICON_MAP duplication
