+++ packages/desktop/src/renderer/components/LeftRail.tsx
@@ -1,19 +1,8 @@
 import React from 'react';
-import { Settings, HelpCircle, MessageSquare, SquareCheck, type LucideProps } from 'lucide-react';
+import { Settings, HelpCircle, MessageSquare } from 'lucide-react';
 import { cn } from '../lib/utils';
 import { usePluginLayoutStore, useSettingsStore } from '../store';
-
-// Curated map of Lucide icon names plugins may declare.
-// Add entries as new plugins are introduced.
-const ICON_MAP: Record<string, React.ComponentType<LucideProps>> = {
-  'square-check': SquareCheck,
-  'message-square': MessageSquare,
-};
-
-function PluginIcon({ name, size = 16 }: { name: string; size?: number }): React.ReactElement | null {
-  const Icon = ICON_MAP[name];
-  return Icon ? <Icon size={size} /> : null;
-}
+import { PluginIcon } from './plugins/PluginIcon';
 
 interface RailButtonProps {
   active?: boolean;
